Expose `require("install").getRequiredIDs` to external code.
This allows packaging tools to scan for requirements in exactly the same way
that install.js will scan for them on the client. Matching false positives is
important so that install.js doesn't accidentally delay the evaluation of a
module that it wrongfully imagines to depend on a nonexistent module.