Support symbolic links.
If the value of an installed filename property is a string, then the
exports object of that file will be identical to the exports object of the
file obtained by resolving the string as a module identifier (relative to
the original file).

This will be especially useful for implementing module substitution in the
spirit of the "browser" field of package.json files.