Expose require.extensions.
This change exposes a list of file extensions that will be tried as part
of module resolution. By default the list is [".js", ".json"], but the
list can be configured via install(tree, { extensions: [...] }).

Note that require.extensions is just a copy of the internal list, so
modifying it won't change any behavior; and it's also an array instead
of an object, unlike require.extensions in Node, which is an object.